Rename type definitions file
The project is called github-workflows so the file name can just be "types" in this context.
Switch from `.d.ts` to `.ts` so it gets included by the typescript compiler
