+++ package.json
@@ -4,7 +4,7 @@
   "description": "Automation to keep Github Actions workflows up to date for Pulumi & third-party repositories",
   "main": "dist/index.js",
   "scripts": {
-    "discovery": "wget https://json.schemastore.org/github-workflow.json -O github-workflow.json && json2ts -i github-workflow.json -o src/github-workflow.d.ts --unreachableDefinitions --ignoreMinAndMaxItems",
+    "discovery": "wget https://json.schemastore.org/github-workflow.json -O github-workflow.json && json2ts -i github-workflow.json -o src/types.ts --unreachableDefinitions --ignoreMinAndMaxItems",
     "test": "echo \"Error: no test specified\" && exit 1",
     "prepack": "rm -rf dist && tsc"
   },
