+++ tsconfig.json
@@ -93,6 +93,6 @@
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
   "include": [
-    "src/*.ts"
+    "src/"
   ]
 }
